+++ packages/cli/__tests__/config/auto-init/__unit__/ApimlAutoInitHandler.unit.test.ts
@@ -141,6 +141,55 @@ describe("ApimlAutoInitHandler", () => {
         expect(response.profiles.base.properties.tokenValue).toEqual("fakeToken");
     });
 
+    it("should not have changed - PEM Certificates", async () => {
+        const mockCreateZosmfSession = jest.fn();
+        const mockGetPluginApimlConfigs = jest.fn().mockReturnValue([]);
+        const mockGetServicesByConfig = jest.fn().mockResolvedValue([]);
+        const mockConvertApimlProfileInfoToProfileConfig = jest.fn().mockReturnValue({
+            defaults: {},
+            profiles: {},
+            plugins: []
+        });
+        const mockLogin = jest.fn().mockResolvedValue("fakeToken");
+
+        ZosmfSession.createSessCfgFromArgs = mockCreateZosmfSession;
+        Services.getPluginApimlConfigs = mockGetPluginApimlConfigs;
+        Services.getServicesByConfig = mockGetServicesByConfig;
+        Services.convertApimlProfileInfoToProfileConfig = mockConvertApimlProfileInfoToProfileConfig;
+        Login.apimlLogin = mockLogin;
+        jest.spyOn(ImperativeConfig.instance, "config", "get").mockReturnValue(mockConfigApi(undefined));
+
+        const handler: any = new ApimlAutoInitHandler();
+        expect(handler.mProfileType).toBe("base");
+
+        handler.createSessCfgFromArgs();
+        expect(mockCreateZosmfSession).toHaveBeenCalledTimes(1);
+
+        const response = await handler.doAutoInit(
+            {
+                ISession: {
+                    hostname: "fake",
+                    port: 1234,
+                    cert: "/fake/cert/file.pem",
+                    certKey: "/fake/cert/key.pem",
+                    type: SessConstants.AUTH_TYPE_CERT_PEM,
+                    tokenType: undefined,
+                }
+            }, {
+                arguments: {
+                    $0: "fake",
+                    _: ["fake"]
+                }
+            });
+        expect(mockGetPluginApimlConfigs).toHaveBeenCalledTimes(1);
+        expect(mockGetServicesByConfig).toHaveBeenCalledTimes(1);
+        expect(mockConvertApimlProfileInfoToProfileConfig).toHaveBeenCalledTimes(1);
+        expect(mockLogin).toHaveBeenCalledTimes(1);
+        expect(response.profiles.base.secure).toContain("tokenValue");
+        expect(response.profiles.base.properties.tokenType).toEqual(SessConstants.TOKEN_TYPE_APIML);
+        expect(response.profiles.base.properties.tokenValue).toEqual("fakeToken");
+    });
+
     it("should not have changed - user & password with existing base profile", async () => {
         const mockCreateZosmfSession = jest.fn();
         const mockGetPluginApimlConfigs = jest.fn().mockReturnValue([]);
@@ -193,7 +242,7 @@ describe("ApimlAutoInitHandler", () => {
         expect(mockGetPluginApimlConfigs).toHaveBeenCalledTimes(1);
         expect(mockGetServicesByConfig).toHaveBeenCalledTimes(1);
         expect(mockConvertApimlProfileInfoToProfileConfig).toHaveBeenCalledTimes(1);
-        expect(mockLogin).toHaveBeenCalledTimes(0);
+        expect(mockLogin).toHaveBeenCalledTimes(1);
         expect(response.profiles.base.secure).not.toContain("tokenValue");
         expect(response.profiles.base.properties.tokenType).not.toBeDefined();
         expect(response.profiles.base.properties.tokenValue).not.toBeDefined();
@@ -398,7 +447,7 @@ describe("ApimlAutoInitHandler", () => {
         expect(mockGetPluginApimlConfigs).toHaveBeenCalledTimes(1);
         expect(mockGetServicesByConfig).toHaveBeenCalledTimes(1);
         expect(mockConvertApimlProfileInfoToProfileConfig).toHaveBeenCalledTimes(0);
-        expect(mockLogin).toHaveBeenCalledTimes(0);
+        expect(mockLogin).toHaveBeenCalledTimes(1);
         expect(error).toBeDefined();
         expect(error.message).toContain("HTTP(S) error status 403 received. Verify the user has access to the APIML API Services SAF resource.");
     });